pre-fix exp3 and exp4

+++ experiments/exp3_feature_analysis.py
@@ -89,7 +89,7 @@ def analyze_feature_activations(clt, model, tokenizer, features: list, prompts: 
 
             acts = clt.encode(sae_input)
             # Get activations for this specific feature at this layer
-            feature_acts = acts[0, :, layer, feature].cpu().numpy()  # [seq_len]
+            feature_acts = acts[0, :, layer, feature].float().cpu().numpy()  # [seq_len]
 
             tokens = tokenizer.convert_ids_to_tokens(inputs[0].tolist())
